+++ src/main/java/open/commons/spring/web/ac/AuthorizedResponse.java
@@ -33,6 +33,10 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.springframework.core.annotation.AliasFor;
+
+import open.commons.spring.web.beans.AbstractResponseDataHandler;
+
 /**
  * 메소드에서 반환하는 데이터를 사용하기 위한 권한을 정의하는 클래스.
  * 
@@ -45,5 +49,49 @@ import java.lang.annotation.Target;
 @Target({ ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface AuthorizedResponse {
+
+    /**
+     * 데이터를 처리하는 Bean 식별정보를 설정합니다. <br>
+     * 설정되는 Bean은 반드시 {@link AbstractResponseDataHandler} 를 구현해야 합니다.
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 5. 20.		박준홍			최초 작성
+     * </pre>
+     *
+     * @return
+     *
+     * @since 2025. 5. 20.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     * 
+     * @see AbstractResponseDataHandler
+     */
+    @AliasFor("value")
+    String dataHandleBean() default "";
+
+    /**
+     * 데이터를 처리하는 Bean 식별정보를 설정합니다. <br>
+     * 설정되는 Bean은 반드시 {@link AbstractResponseDataHandler} 를 구현해야 합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 5. 20.		박준홍			최초 작성
+     * </pre>
+     *
+     * @return
+     *
+     * @since 2025. 5. 20.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     * 
+     * @see AbstractResponseDataHandler
+     */
+    @AliasFor("dataHandleBean")
+    String value() default "";
 
 }