+++ src/main/java/open/commons/spring/web/aspect/AuthorizedMethodAspect.java
@@ -105,11 +105,15 @@ public class AuthorizedMethodAspect extends AbstractAuthorizedResourceAspect<IMe
 
         // 타입과 메소드에 모두 AuthorizedMethod 가 설정된 경우 메소드에 설정된 어노테이션을 사용함.
         AuthorizedMethod annotation = decideAnnotation(AuthorizedMethod.class, target.getClass(), method);
-        String beanName = annotation.bean();
-        IMethodAccessAuthorityProvider bean = getBean(beanName);
+        String beanName = annotation.authorityBean();
+        IMethodAccessAuthorityProvider bean = getAuthorityBean(beanName);
 
         logger.trace("annotation={}", annotation);
         logger.trace("provider={}", bean);
+
+        if (annotation.roles().length < 1) {
+            throw new UnauthorizedException("올바르지 않은 접근입니다.");
+        }
 
         Result<Boolean> validated = bean.isAllowed(annotation.op(), annotation.roles());
         if (!validated.getResult() || !validated.getData()) {