+++ src/main/java/open/commons/spring/web/aspect/AuthorizedResponseAspect.java
@@ -27,8 +27,8 @@
 package open.commons.spring.web.aspect;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
@@ -37,15 +37,21 @@ import javax.validation.constraints.NotNull;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.reflect.MethodSignature;
 import org.springframework.context.ApplicationContext;
+import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.core.annotation.Order;
 
 import open.commons.core.Result;
-import open.commons.core.utils.AnnotationUtils;
 import open.commons.core.utils.ExceptionUtils;
-import open.commons.core.utils.ObjectUtils;
+import open.commons.core.utils.ReflectionUtils;
 import open.commons.spring.web.ac.AuthorizedField;
-import open.commons.spring.web.ac.provider.IResponseAccessAuthorityProvider;
+import open.commons.spring.web.ac.AuthorizedField.AuthorizedFieldModeHandle;
+import open.commons.spring.web.ac.AuthorizedResponse;
+import open.commons.spring.web.ac.provider.IFieldAccessAuthorityProvider;
+import open.commons.spring.web.beans.AbstractResponseDataHandler;
+import open.commons.spring.web.beans.DefaultUnauthorizedFieldHandler;
+import open.commons.spring.web.beans.IUnauthorizedFieldHandler;
 import open.commons.spring.web.servlet.InternalServerException;
 import open.commons.spring.web.servlet.UnauthorizedException;
 
@@ -57,7 +63,7 @@ import open.commons.spring.web.servlet.UnauthorizedException;
  */
 @Aspect
 @Order(AbstractAuthorizedResourceAspect.ORDER_RESPONSE)
-public class AuthorizedResponseAspect extends AbstractAuthorizedResourceAspect<IResponseAccessAuthorityProvider> {
+public class AuthorizedResponseAspect extends AbstractAuthorizedResourceAspect<IFieldAccessAuthorityProvider> {
 
     /**
      * <br>
@@ -75,11 +81,13 @@ public class AuthorizedResponseAspect extends AbstractAuthorizedResourceAspect<I
      * @author parkjunhong77@gmail.com
      */
     public AuthorizedResponseAspect(@NotNull ApplicationContext context) {
-        super(context, IResponseAccessAuthorityProvider.class);
+        super(context, IFieldAccessAuthorityProvider.class);
     }
 
     @Around("withinAllStereotypeComponent() && annotationAuthorizedResponse() ")
     public Object validateAuthorizedMethod(ProceedingJoinPoint pjp) throws Throwable {
+
+        Method method = ((MethodSignature) pjp.getSignature()).getMethod();
 
         // #1. 메소드 결과괎
         Object result = pjp.proceed();
@@ -87,34 +95,40 @@ public class AuthorizedResponseAspect extends AbstractAuthorizedResourceAspect<I
         logger.trace("result.type={}", result != null ? result.getClass() : null);
         logger.trace("result={}", result);
 
-        // #2. 지원하는 Wrapper 클래스
-        // - java.util.Collection
-        // - java.util.List
-        // - java.util.Map
-        // - java.util.Set
-        // - open.commons.core.Result
-        Class<?> outerClass = result.getClass();
-        if (outerClass.isAssignableFrom(Collection.class)) {
-            Collection<?> col = (Collection<?>) result;
-            for (Object o : col) {
-                validateObject(o);
-            }
-        } else if (outerClass.isAssignableFrom(Map.class)) {
-            Map<?, ?> m = (Map<?, ?>) result;
-            for (Entry<?, ?> e : m.entrySet()) {
-                validateObject(e.getValue());
-            }
-        } else if (outerClass.isAssignableFrom(Result.class)) {
-            Result<?> r = (Result<?>) result;
-            validateObject(r.getData());
+        // #2. 메소드에 설정된 어노테이션
+        AuthorizedResponse annoRes = AnnotationUtils.getAnnotation(method, AuthorizedResponse.class);
+        // 데이터를 처리하는 Bean
+        String beanName = annoRes.dataHandleBean();
+        if (!beanName.isEmpty()) {
+            AbstractResponseDataHandler handler = this.context.getBean(beanName, AbstractResponseDataHandler.class);
+            result = handler.handle(result);
         } else {
-            validateObject(result);
+
+            // #2. 지원하는 Wrapper 클래스
+            // - java.util.Collection
+            // - java.util.List
+            // - java.util.Map
+            // - java.util.Set
+            // - open.commons.core.Result
+            Class<?> outerClass = result.getClass();
+            if (outerClass.isAssignableFrom(Collection.class)) {
+                Collection<?> col = (Collection<?>) result;
+                for (Object o : col) {
+                    validateObject(o);
+                }
+            } else if (outerClass.isAssignableFrom(Map.class)) {
+                Map<?, ?> m = (Map<?, ?>) result;
+                for (Entry<?, ?> e : m.entrySet()) {
+                    validateObject(e.getValue());
+                }
+            } else if (outerClass.isAssignableFrom(Result.class)) {
+                Result<?> r = (Result<?>) result;
+                validateObject(r.getData());
+            } else {
+                validateObject(result);
+            }
         }
-
         return result;
-    }
-
-    private void nullfy(Object o, Field f) {
     }
 
     private void validateObject(Object o) throws IllegalArgumentException, IllegalAccessException {
@@ -122,36 +136,48 @@ public class AuthorizedResponseAspect extends AbstractAuthorizedResourceAspect<I
             return;
         }
 
-        // TODO: 여기서부터 클래스 필드를 하나씩 조사하면서 데이터 검증
-        List<Field> fields = AnnotationUtils.getAnnotatedFieldsAllHierarchy(o.getClass(), AuthorizedField.class);
-        AuthorizedField anno = null;
-        IResponseAccessAuthorityProvider bean = null;
-        Result<Boolean> validated = null;
-        Object newValue = null;
-        for (Field f : fields) {
-            anno = f.getAnnotation(AuthorizedField.class);
-            bean = getBean(anno.bean());
+        // #1. 타입 수준에서 정의한 권한제한 필드 처리 조회
+        AuthorizedFieldModeHandle annoModeHandle = AnnotationUtils.findAnnotation(o.getClass(), AuthorizedFieldModeHandle.class);
+        IUnauthorizedFieldHandler typeModeHandle = null;
+        if (annoModeHandle != null) {
+            typeModeHandle = getBean(annoModeHandle.bean(), DefaultUnauthorizedFieldHandler.class, null, true);
+        }
 
-            validated = bean.isAllowed(anno.op(), anno.roles());
+        // #2. 모든 필드 조회
+        Map<Field, AuthorizedField> fields = ReflectionUtils.getAllAnnotatedFields(o, AuthorizedField.class);
+
+        Field field = null;
+        AuthorizedField fieldAnno = null;
+        IFieldAccessAuthorityProvider fieldAuthorityBean = null;
+        Result<Boolean> validated = null;
+        IUnauthorizedFieldHandler fieldModeHandle = null;
+
+        for (Entry<Field, AuthorizedField> entry : fields.entrySet()) {
+            field = entry.getKey();
+            fieldAnno = entry.getValue();
+            fieldAuthorityBean = getAuthorityBean(fieldAnno.authorityBean());
+
+            if (fieldAnno.roles().length < 1) {
+                throw new UnauthorizedException("올바르지 않은 접근입니다.");
+            }
+
+            validated = fieldAuthorityBean.isAllowed(fieldAnno.op(), fieldAnno.roles());
             if (!validated.getResult()) {
                 throw ExceptionUtils.newException(InternalServerException.class, "데이터 접근권한 처리 중 오류가 발생하였습니다. 원인=%s", validated.getMessage());
             } else if (!validated.getData()) {
-                switch (anno.mode()) {
+                fieldModeHandle = getBean(fieldAnno.modeHandleBean(), IUnauthorizedFieldHandler.class, typeModeHandle, false);
+                switch (fieldAnno.mode()) {
                     case DENY:
-                        throw new UnauthorizedException("올바르지 않은 접근입니다.");
+                        fieldModeHandle.deny(o, field);
+                        break;
                     case MASK:
-                        // TODO: Masking 적용 대상 확인
-                        boolean assessible = f.isAccessible();
-                        f.setAccessible(true);
-                        newValue = bean.mask(o, f, "", (String) f.get(o));
-                        f.set(o, newValue);
-                        f.setAccessible(assessible);
+                        fieldModeHandle.mask(o, field, fieldAnno.masking());
                         break;
                     case NULLIFY:
-                        nullfy(o, f);
+                        fieldModeHandle.nullify(o, field);
                         break;
                     default:
-                        throw ExceptionUtils.newException(UnsupportedOperationException.class, "지원하지 않는 기능입니다. 원인=%s", anno.mode());
+                        throw ExceptionUtils.newException(UnsupportedOperationException.class, "지원하지 않는 기능입니다. 원인=%s", fieldAnno.mode());
                 }
             }
         }